+++ libft/ft_bzero.c
@@ -1,0 +1,25 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_bzero.c                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vnafissi <vnafissi@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/04/11 18:08:49 by vnafissi          #+#    #+#             */
+/*   Updated: 2022/04/11 18:09:07 by vnafissi         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "libft.h"
+
+void	ft_bzero(void *ptr, size_t n)
+{
+	char	*s;
+
+	s = (char *) ptr;
+	while (n > 0)
+	{
+		s[n - 1] = 0;
+		n--;
+	}
+}
