+++ libft/ft_calloc.c
@@ -1,18 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   libft.c                                            :+:      :+:    :+:   */
+/*   ft_calloc.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vnafissi <vnafissi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/04/11 12:19:09 by vnafissi          #+#    #+#             */
-/*   Updated: 2022/04/11 16:03:17 by vnafissi         ###   ########.fr       */
+/*   Created: 2022/04/11 18:08:19 by vnafissi          #+#    #+#             */
+/*   Updated: 2022/04/11 18:08:40 by vnafissi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minishell.h"
-
-
+#include "libft.h"
 
 // allocates memory for an array of n elements of size bytes each and returns
 //  a pointer to the allocated memory.
@@ -27,44 +25,3 @@ void	*ft_calloc(size_t n, size_t size)
 	return (ptr);
 }
 
-void	ft_bzero(void *ptr, size_t n)
-{
-	char	*s;
-
-	s = (char *) ptr;
-	while (n > 0)
-	{
-		s[n - 1] = 0;
-		n--;
-	}
-}
-
-char	*ft_strdup(const char *s, int len)
-{
-	char	*dup;
-	int		i;
-
-	dup = ft_calloc(len + 1 , sizeof(char));
-	if (!dup)
-		return (NULL);
-	i = 0;
-	while (s[i])
-	{
-		dup[i] = s[i];
-		i++;
-	}
-	dup[i] = '\0';
-	return (dup);
-}
-
-// calcule la longueur de la chaîne de caractères s,
-// sans compter l'octet nul « \0 » final.
-size_t	ft_strlen(const char *s)
-{
-	size_t	i;
-
-	i = 0;
-	while (s[i])
-		i++;
-	return (i);
-}
